+++ packages/boxed-expression-component/package.json
@@ -22,7 +22,7 @@
     "@kie-tools-core/monaco-editor": "workspace:*",
     "@kie-tools-core/operating-system": "workspace:*",
     "@kie-tools-core/patternfly-base": "workspace:*",
-    "@kie-tools/dmn-language-service": "workspace:*",
+    "@kie-tools/dmn-feel-antlr4-parser": "workspace:*",
     "@kie-tools/feel-input-component": "workspace:*",
     "@kie-tools/i18n-common-dictionary": "workspace:*",
     "@patternfly/react-core": "^4.276.6",
