+++ packages/dmn-feel-antlr4-parser/src/index.ts
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *  http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -17,23 +17,9 @@
  * under the License.
  */
 
-import { FeelVariablesParser } from "./parser/FeelVariablesParser";
-import { VariablesRepository } from "./parser/VariablesRepository";
-
-export class FeelVariables {
-  private readonly _parser: FeelVariablesParser;
-  private readonly _repository: VariablesRepository;
-
-  constructor(xml: string) {
-    this._repository = new VariablesRepository(xml);
-    this._parser = new FeelVariablesParser(this._repository);
-  }
-
-  get parser(): FeelVariablesParser {
-    return this._parser;
-  }
-
-  get repository(): VariablesRepository {
-    return this._repository;
-  }
-}
+export * from "./FeelVariables";
+export * from "./parser/FeelVariable";
+export * from "./parser/ReservedWords";
+export * from "./parser/VariableType";
+export * from "./parser/FeelVariablesParser";
+export * from "./parser/VariablesRepository";
