+++ packages/dmn-feel-antlr4-parser/src/parser/VariableType.ts
