+++ packages/dmn-feel-antlr4-parser/src/parser/grammar/ScopeImpl.ts
