+++ packages/dmn-feel-antlr4-parser/src/parser/grammar/Scopes.ts
