+++ packages/dmn-feel-antlr4-parser/src/parser/grammar/VariableSymbol.ts
