+++ packages/dmn-feel-antlr4-parser/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@kie-tools/tsconfig/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src"]
+}
